requirements.txt
alabaster==0.7.10
anaconda-client==1.6.3
anaconda-navigator==1.6.2
anaconda-project==0.6.0
arrow==0.12.1
asn1crypto==0.22.0
astroid==1.4.9
astropy==1.3.2
attrs==17.2.0
Automat==0.0.0
Babel==2.4.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
binaryornot==0.4.4
bitarray==0.8.1
blaze==0.10.1
bleach==1.5.0
bokeh==0.12.5
boto==2.46.1
Bottleneck==1.2.1
certifi==2018.1.18
cffi==1.10.0
chardet==3.0.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
colorama==0.3.9
conda==4.3.30
constantly==15.1.0
contextlib2==0.5.5
cookiecutter==1.6.0
cryptography==1.8.1
cssselect==1.0.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.14.3
datashape==0.5.4
decorator==4.0.11
distributed==1.16.3
Django==1.11.6
docutils==0.13.1
-e git+https://github.com/Ematrix163/Dublin_bikes.git@5dae34f14ef02e2c6dbd4fb98b8f1c06afd84f19#egg=dublin_bikes
entrypoints==0.2.2
et-xmlfile==1.0.1
fastcache==1.0.2
feedparser==5.2.1
Flask==0.12.2
Flask-Cors==3.0.2
future==0.16.0
gevent==1.2.1
greenlet==0.4.12
h5py==2.7.0
HeapDict==1.0.0
html5lib==0.9999999
hyperlink==17.3.1
hypothesis==3.33.0
idna==2.5
imagesize==0.7.1
incremental==17.5.0
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jinja2-time==0.2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.2.2
-e git+https://github.com/diarmuidmorgan/lightz.git@5a3463a08b72a2c4d75bf20f3782ac0be426e83a#egg=lightz
llvmlite==0.18.0
locket==0.2.0
lxml==3.7.3
Markdown==2.6.9
MarkupSafe==0.23
matplotlib==2.0.2
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
mysql-connector-python==2.0.4
mysqlclient==1.3.12
navigator-updater==0.1.0
nbconvert==5.1.1
nbformat==4.3.0
networkx==1.11
nltk==3.2.3
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.12.1
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
openpyxl==2.4.7
packaging==16.8
pandas==0.20.1
pandocfilters==1.4.1
parsel==1.2.0
partd==0.3.8
pathlib2==2.2.1
patsy==0.4.1
pdfminer.six==20170720
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
ply==3.10
poyo==0.4.1
prompt-toolkit==1.0.14
protobuf==3.4.0
psutil==5.2.2
ptyprocess==0.5.1
py==1.4.33
pyasn1==0.3.7
pyasn1-modules==0.1.4
pycosat==0.6.2
pycparser==2.17
pycrypto==2.6.1
pycryptodome==3.4.8
pycurl==7.43.0
PyDispatcher==2.0.5
pyflakes==1.5.0
Pygments==2.2.0
pylint==1.6.4
Pympler==0.5
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyparsing==2.1.4
## !! Could not determine repository location
pyscaf-project==0.0.post0.dev1+g2181660
PyScaffold==3.0.1
pytest==3.0.7
python-dateutil==2.6.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
queuelib==1.4.2
requests==2.18.4
rope-py3k==0.9.4.post1
scikit-image==0.13.0
scikit-learn==0.18.1
scipy==0.19.0
Scrapy==1.4.0
seaborn==0.7.1
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.5.6
spyder==3.1.4
SQLAlchemy==1.1.9
statsmodels==0.8.0
sudoku-maker==1.0.12
sympy==1.0
-e git+https://github.com/diarmuidmorgan/systeminfo.git@9a5f4eef112ddcde53b510c2dff92b708bebabba#egg=systeminfo
tables==3.3.0
tblib==1.3.2
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
terminado==0.6
testpath==0.3
toolz==0.8.2
tornado==4.5.1
traitlets==4.3.2
Twisted==17.5.0
twitter==1.18.0
unicodecsv==0.14.1
urllib3==1.22
w3lib==1.17.0
wcwidth==0.1.7
Werkzeug==0.12.2
whichcraft==0.4.1
widgetsnbextension==2.0.0
wrapt==1.10.10
xlrd==1.0.0
XlsxWriter==0.9.6
xlwt==1.2.0
zict==0.1.2
zope.interface==4.4.2
